+++ docs/docs-generator.js
@@ -144,9 +144,9 @@ Comment.concatTypes = function (types, convertEntities) {
   if (convertEntities) {
     // Convert a couple of things to their HTML-entities
     // The spacing around | is intentional, in order to introduce some linebreaks in the params table
-    type = type.replace('|', ' &#124; ')
-      .replace('>', '&gt;')
-      .replace('<', '&lt;');
+    type = type.replace(/\|/g, ' &#124; ')
+      .replace(/>/g, '&gt;')
+      .replace(/</g, '&lt;');
   }
 
   return type;
@@ -213,7 +213,7 @@ var parseComments = function (comments, file) {
           if (see.local) {
             link = see.local.match(/{(.*?(?:|#.*?))}/)[1];
 
-            comment.putLine('* [' + link + '](' + link.toLowerCase() + ')');
+            comment.putLine('* [' + link + '](api/' + link.toLowerCase() + ')');
           } else {
             comment.putLine('* [' + see.title | see.url + '](' + see.url + ')');
           }