+++ test/integration/model/upsert.test.js
@@ -18,8 +18,6 @@ chai.config.includeStack = true;
 
 describe(Support.getTestDialectTeaser('Model'), function() {
   beforeEach(function() {
-    this.clock = sinon.useFakeTimers();
-
     this.User = this.sequelize.define('user', {
       username: DataTypes.STRING,
       foo: {
@@ -29,14 +27,15 @@ describe(Support.getTestDialectTeaser('Model'), function() {
       bar: {
         unique: 'foobar',
         type: DataTypes.INTEGER
-      }
+      },
+      baz: {
+        type: DataTypes.STRING,
+        field: 'zab'
+      },
+      blob: DataTypes.BLOB
     });
 
     return this.sequelize.sync({ force: true });
-  });
-
-  afterEach(function() {
-    this.clock.restore();
   });
 
   if (current.dialect.supports.upserts) {
@@ -49,8 +48,9 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          this.clock.tick(2000); // Make sure to pass some time so updatedAt != createdAt
-          return this.User.upsert({ id: 42, username: 'doe' });
+          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
+            return this.User.upsert({ id: 42, username: 'doe' });
+          });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).not.to.be.defined;
@@ -74,8 +74,9 @@ describe(Support.getTestDialectTeaser('Model'), function() {
             expect(created).to.be.ok;
           }
 
-          this.clock.tick(2000); // Make sure to pass some time so updatedAt != createdAt
-          return this.User.upsert({ foo: 'baz', bar: 19, username: 'doe' });
+          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
+            return this.User.upsert({ foo: 'baz', bar: 19, username: 'doe' });
+          });
         }).then(function(created) {
           if (dialect === 'sqlite') {
             expect(created).not.to.be.defined;
@@ -91,6 +92,60 @@ describe(Support.getTestDialectTeaser('Model'), function() {
         });
       });
 
+      it('works with upsert on a composite primary key', function() {
+        var User = this.sequelize.define('user', {
+          a: {
+            type: Sequelize.STRING,
+            primaryKey: true,
+          },
+          b: {
+            type: Sequelize.STRING,
+            primaryKey: true,
+          },
+          username: DataTypes.STRING,
+        });
+
+        return User.sync({ force: true }).bind(this).then(function  () {
+          return Promise.all([
+              // Create two users
+             User.upsert({ a: 'a', b: 'b', username: 'john' }),
+             User.upsert({ a: 'a', b: 'a', username: 'curt' }),
+          ]);
+        }).spread(function(created1, created2) {
+          if (dialect === 'sqlite') {
+            expect(created1).not.to.be.defined;
+            expect(created2).not.to.be.defined;
+          } else {
+            expect(created1).to.be.ok;
+            expect(created2).to.be.ok;
+          }
+
+          return Promise.delay(1000).bind(this).then(function() {
+            // Update the first one
+            return User.upsert({ a: 'a', b: 'b', username: 'doe' });
+          });
+        }).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).not.to.be.defined;
+          } else {
+            expect(created).not.to.be.ok;
+          }
+
+          return User.find({ where: { a: 'a', b: 'b' }});
+        }).then(function (user1) {
+          expect(user1.createdAt).to.be.defined;
+          expect(user1.username).to.equal('doe');
+          expect(user1.updatedAt).to.be.afterTime(user1.createdAt);
+
+          return User.find({ where: { a: 'a', b: 'a' }});
+        }).then(function (user2) {
+          // The second one should not be updated
+          expect(user2.createdAt).to.be.defined;
+          expect(user2.username).to.equal('curt');
+          expect(user2.updatedAt).to.equalTime(user2.createdAt);
+        });
+      });
+
       it('supports validations', function () {
         var User = this.sequelize.define('user', {
           email: {
@@ -102,6 +157,57 @@ describe(Support.getTestDialectTeaser('Model'), function() {
         });
 
         return expect(User.upsert({ email: 'notanemail' })).to.eventually.be.rejectedWith(this.sequelize.ValidationError);
+      });
+
+      it('works with BLOBs', function () {
+        return this.User.upsert({ id: 42, username: 'john', blob: new Buffer('kaj') }).bind(this).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).not.to.be.defined;
+          } else {
+            expect(created).to.be.ok;
+          }
+
+          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
+            return this.User.upsert({ id: 42, username: 'doe', blob: new Buffer('andrea') });
+          });
+        }).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).not.to.be.defined;
+          } else {
+            expect(created).not.to.be.ok;
+          }
+
+          return this.User.find(42);
+        }).then(function(user) {
+          expect(user.createdAt).to.be.defined;
+          expect(user.username).to.equal('doe');
+          expect(user.blob.toString()).to.equal('andrea');
+          expect(user.updatedAt).to.be.afterTime(user.createdAt);
+        });
+      });
+
+      it('works with .field', function () {
+        return this.User.upsert({ id: 42, baz: 'foo' }).bind(this).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).not.to.be.defined;
+          } else {
+            expect(created).to.be.ok;
+          }
+
+          return this.sequelize.Promise.delay(1000).bind(this).then(function() {
+            return this.User.upsert({ id: 42, baz: 'oof' });
+          });
+        }).then(function(created) {
+          if (dialect === 'sqlite') {
+            expect(created).not.to.be.defined;
+          } else {
+            expect(created).not.to.be.ok;
+          }
+
+          return this.User.find(42);
+        }).then(function(user) {
+          expect(user.baz).to.equal('oof');
+        });
       });
     });
   }